backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
openpyxl==3.1.2
xlrd==2.0.1
pypdf==3.17.4
python-docx==1.1.0
pandas>=2.2.0
numpy==1.26.2
python-dotenv==1.0.0

